+++ example/macos/WGSystemExtension/main.swift
@@ -1,0 +1,15 @@
+//
+//  main.swift
+//  WGSystemExtension
+//
+//  Created by Akash Chandrakar on 12/05/26.
+//
+
+import Foundation
+import NetworkExtension
+
+autoreleasepool {
+    NEProvider.startSystemExtensionMode()
+}
+
+dispatchMain()